requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\heh\NN-arena\src\Game\playbestneat.py: 4
# D:\heh\NN-arena\src\Game\train.py: 3
neat_python == 0.92

# D:\heh\NN-arena\src\Game\main.py: 1
# D:\heh\NN-arena\src\Game\player.py: 2
# D:\heh\NN-arena\src\Game\projectile.py: 1
# D:\heh\NN-arena\src\Game\senses.py: 1
pygame

# D:\heh\NN-arena\src\Game\train.py: 5
tqdm
